requirements.txt
PyMuPDF>=1.23.0
openai>=1.0.0
chromadb>=0.4.0
python-dotenv>=1.0.0
streamlit>=1.28.0

